+++ src/main/preload.ts
@@ -20,14 +20,17 @@ export type Channels =
   | 'narrative-changed'
   | 'export-full-manuscript'
   | 'world-objects-changed'
+  | 'templates-changed'
   | 'open-template-manager'
   | 'open-project-settings'
   | 'show-notification'
   | 'open-project-wizard'
+  | 'open-import-from-file-modal'
   | 'item-deleted'
   | 'project-archive'
   | 'project:archive-request'
-  | 'trigger-export-world-objects';
+  | 'trigger-export-world-objects'
+  | 'trigger-import-world-objects';
 
 const electronHandler = {
   ipcRenderer: {
@@ -97,6 +100,12 @@ export type ElectronAPI = typeof electronHandler & {
       error?: string;
       canceled?: boolean;
     }>;
+    import: () => Promise<{
+      success: boolean;
+      filePath?: string;
+      error?: string;
+      canceled?: boolean;
+    }>;
   };
   fs: {
     openFolder: (folderPath: string) => Promise<void>;
@@ -131,6 +140,7 @@ contextBridge.exposeInMainWorld('electron', {
   },
   worldObjects: {
     export: () => ipcRenderer.invoke('export:world-objects'),
+    import: () => ipcRenderer.invoke('trigger-import-world-objects'),
   },
   fs: {
     openFolder: (folderPath: string) =>